pyproject.toml
[tool.poetry]
name = "managerngrok"
version = "1.0.0"
description = ""
authors = ["GutesTod <jefflol003@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
